Page/playlist loading bugfix.
media.i keeps track of which song is playing.  It wasn't being updated when a
playlist was loaded as part of the page load (i.e., via GET params).  The result
was that "next" and "previous" functionality was acting as though the first song
in the playlist was playing, instead of the one that actually was.